% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapeau.R
\name{bt_error}
\alias{bt_error}
\title{Title}
\usage{
bt_error(X, Y, type, Sigma_12inv = diag(1, ncol(as.data.frame(Y))),
  group = NULL, a = 1, times = 10, sep = "\\\\.")
}
\arguments{
\item{a}{}
}
\description{
Title
}
