% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapeau.R
\name{compar_type}
\alias{compar_type}
\title{Title}
\usage{
compar_type(X = NULL, Y, types, Sigma_12inv = NULL, group = NULL,
  a = 1, times = 10, sep = "\\\\.", regressors = NULL,
  type_S12_inv = "emp")
}
\arguments{
\item{times}{}
}
\description{
Title
}
