% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_lasso.R
\name{fused_lasso}
\alias{fused_lasso}
\title{Title}
\usage{
fused_lasso(
  X,
  response,
  G,
  lambda2 = NULL,
  lambda1 = 0,
  ratio = 0.001,
  nlambda2 = 100
)
}
\arguments{
\item{fixe_lambda1}{}
}
\value{
fused lasso model
}
\description{
Title
}
