% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_lasso.R
\name{get_group_both}
\alias{get_group_both}
\title{Description of the function}
\usage{
get_group_both(name_x, sep = "\\\\.", r)
}
\arguments{
\item{sep}{the separators between the group (for exemlpe the marker) and the specificity of the columns (for exemple the allele)}

\item{r}{the number of Trait}

\item{X}{a matrix with colnames indicating group vaiables}
}
\value{
The group for r Trait if the columns have to be group if they have the same Marker, it is the same group for all Trait
}
\description{
Description of the function
}
\examples{
B <- c(1, -1, 1.5, 1.5, rep(0, 6), 2, 0, 2, 0)
group <- c(rep('M1', 10), rep('M2', 10))
regressors <- matrix(rnorm(6*20), ncol = 6)
X  <- model.matrix(~group + group:regressors - 1)
y <- X\%*\%B + rnorm(20)
y <- scale(y)
mod <- fl2(y, regressors, group)
colors <- c(rep("grey",2), rep('green',2),rep('black', 6), rep(c("orange","blue"), 2), 'darkgreen', rep('yellow',3), rep('purple',2))
matplot(mod$lambda ,t(mod$beta),type='l',col=colors)
}
