% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_lasso.R
\name{get_group_marker}
\alias{get_group_marker}
\title{This function give the group for r Traits if the group are just among the Marker and not among the Trait}
\usage{
get_group_marker(name_x, sep = "\\\\.", r)
}
\description{
This function give the group for r Traits if the group are just among the Marker and not among the Trait
}
