% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_lasso.R
\name{grp_lasso}
\alias{grp_lasso}
\title{Description of the function}
\usage{
grp_lasso(X, y, group, dfmax = as.integer(max(group)) + 1,
  pmax = min(dfmax * 1.2, as.integer(max(group))))
}
\description{
Description of the function
}
