% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_lasso.R
\name{grpLassoQTL}
\alias{grpLassoQTL}
\title{Give the result for a group lasso}
\usage{
grpLassoQTL(
  X,
  Y,
  marker,
  map,
  PFER = 1,
  B = 500,
  cutoff = 0.85,
  nb.cores,
  sep = "\\\\.",
  mrk2lg,
  type_group = "both"
)
}
\arguments{
\item{X}{a matrix with the value of different allele for different Marker}

\item{Y}{a vector or a matrix with the values of different Trait}

\item{type_group}{a character indicating the way of creating the group :
"both" means that the variable are selected for all a Marker for all the Trait.
"marker" means that the variable are selected for a Marker but for one Trait.
"trait" means that a variable are selected for all the Trait but not for all the marker but just for one allele}

\item{a}{the parameters that indicate how much the coefficients will be fused}

\item{lambda}{if the user wants to use it owns values of lambdas}
}
\value{
The coefficients of the fused lasso ANCOVA for the different value of lambda
}
\description{
Give the result for a group lasso
}
\examples{
B <- c(1, -1, 1.5, 1.5, rep(0, 6), 2, 0, 2, 0)
group <- c(rep('M1', 10), rep('M2', 10))
regressors <- matrix(rnorm(6*20), ncol = 6)
X  <- model.matrix(~group + group:regressors - 1)
y <- X\%*\%B + rnorm(20)
y <- scale(y)
mod <- fl2(y, regressors, group)
colors <- c(rep("grey",2), rep('green',2),rep('black', 6), rep(c("orange","blue"), 2), 'darkgreen', rep('yellow',3), rep('purple',2))
matplot(mod$lambda ,t(mod$beta),type='l',col=colors)
}
