% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fus2.R
\docType{data}
\name{mod_fus2_univ}
\alias{mod_fus2_univ}
\title{Title}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
mod_fus2_univ
}
\arguments{
\item{...}{}
}
\description{
Title
}
\keyword{datasets}
