% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_fused.R
\docType{data}
\name{mod_fused_multi_regr}
\alias{mod_fused_multi_regr}
\title{Description of the function}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
mod_fused_multi_regr
}
\description{
Description of the function
}
\keyword{datasets}
