% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_group.R
\docType{data}
\name{mod_group_multi}
\alias{mod_group_multi}
\title{Description of the function}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
mod_group_multi
}
\description{
Description of the function
}
\keyword{datasets}
