% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_lasso.R
\name{mod_lasso}
\alias{mod_lasso}
\title{Description of the function}
\value{
The coefficients of the fused lasso ANCOVA for the different value of lambda
}
\description{
Description of the function

Description of the function
}
\examples{
B <- c(1, -1, 1.5, 1.5, rep(0, 6), 2, 0, 2, 0)
group <- c(rep('M1', 10), rep('M2', 10))
regressors <- matrix(rnorm(6*20), ncol = 6)
X  <- model.matrix(~group + group:regressors - 1)
y <- X\%*\%B + rnorm(20)
y <- scale(y)
mod <- fl2(y, regressors, group)
colors <- c(rep("grey",2), rep('green',2),rep('black', 6), rep(c("orange","blue"), 2), 'darkgreen', rep('yellow',3), rep('purple',2))
matplot(mod$lambda ,t(mod$beta),type='l',col=colors)
#' @import R6 Matrix gglasso tidyverse glmnet stabs magrittr viridis stringr FusedLasso
}
\section{Super class}{
\code{\link[VariSel:VariSel]{VariSel::VariSel}} -> \code{mod_lasso}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{mod_lasso$new()}}
\item \href{#method-estime}{\code{mod_lasso$estime()}}
\item \href{#method-sel_cv}{\code{mod_lasso$sel_cv()}}
\item \href{#method-sel_stabs}{\code{mod_lasso$sel_stabs()}}
\item \href{#method-plot_path}{\code{mod_lasso$plot_path()}}
\item \href{#method-clone}{\code{mod_lasso$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="get_coef">}\href{../../VariSel/html/VariSel.html#method-get_coef}{\code{VariSel::VariSel$get_coef()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="plot_anime">}\href{../../VariSel/html/VariSel.html#method-plot_anime}{\code{VariSel::VariSel$plot_anime()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="plot_BIC">}\href{../../VariSel/html/VariSel.html#method-plot_BIC}{\code{VariSel::VariSel$plot_BIC()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="plot_coef">}\href{../../VariSel/html/VariSel.html#method-plot_coef}{\code{VariSel::VariSel$plot_coef()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="plot_cv">}\href{../../VariSel/html/VariSel.html#method-plot_cv}{\code{VariSel::VariSel$plot_cv()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="plot_error">}\href{../../VariSel/html/VariSel.html#method-plot_error}{\code{VariSel::VariSel$plot_error()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="predict">}\href{../../VariSel/html/VariSel.html#method-predict}{\code{VariSel::VariSel$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="ROC">}\href{../../VariSel/html/VariSel.html#method-ROC}{\code{VariSel::VariSel$ROC()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_lasso$new(
  x,
  y,
  univ = TRUE,
  Sigma_12inv = diag(1, ncol(as.data.frame(y))),
  sepy = "__",
  penalty.factor = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estime"></a>}}
\subsection{Method \code{estime()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_lasso$estime(lambda = NULL, standardize = FALSE, intercept = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lambda}}{if the user wants to use it owns values of lambdas}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sel_cv"></a>}}
\subsection{Method \code{sel_cv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_lasso$sel_cv(s = "lambda.min")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sel_stabs"></a>}}
\subsection{Method \code{sel_stabs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_lasso$sel_stabs(nb.cores = 7, B = 500, PFER = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_path"></a>}}
\subsection{Method \code{plot_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_lasso$plot_path(type = "first", nb = 6)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_lasso$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
