% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ord_beta}
\alias{ord_beta}
\title{Description of the function}
\usage{
ord_beta(beta, name)
}
\arguments{
\item{response}{a vector response variable}

\item{regressors}{a quantitative matrix of regressor}

\item{group}{a vector with two levels. (The group of the ANCOVA)}

\item{a}{the parameters that indicate how much the coefficients will be fused}

\item{lambda}{if the user wants to use it owns values of lambdas}
}
\value{
The coefficients of the fused lasso ANCOVA for the different value of lambda
}
\description{
Description of the function
}
\examples{
B <- c(1, -1, 1.5, 1.5, rep(0, 6), 2, 0, 2, 0)
group <- c(rep('M1', 10), rep('M2', 10))
regressors <- matrix(rnorm(6*20), ncol = 6)
X  <- model.matrix(~group + group:regressors - 1)
y <- X\%*\%B + rnorm(20)
y <- scale(y)
mod <- fl2(y, regressors, group)
colors <- c(rep("grey",2), rep('green',2),rep('black', 6), rep(c("orange","blue"), 2), 'darkgreen', rep('yellow',3), rep('purple',2))
matplot(mod$lambda ,t(mod$beta),type='l',col=colors)
}
