% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_lasso_penalty.R
\docType{data}
\name{QTLmod_lasso_adaptive}
\alias{QTLmod_lasso_adaptive}
\title{Description of the function}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QTLmod_lasso_adaptive
}
\description{
Description of the function
}
\keyword{datasets}
