% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chap_varisel.R
\name{train_VariSel}
\alias{train_VariSel}
\title{Title}
\usage{
train_VariSel(X = NULL, Y, type, sepx = NULL, regressors = NULL,
  group = NULL, lambda = NULL, lambda1 = NULL, lambda2 = NULL,
  Sigma_12inv = NULL, type_S12_inv = "fixed")
}
\arguments{
\item{lambda2}{}
}
\description{
Title
}
