% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTLmodclasss.R
\name{VariSel}
\alias{VariSel}
\title{Description of the function}
\value{
The coefficients of the fused lasso ANCOVA for the different value of lambda
}
\description{
Description of the function

Description of the function
}
\examples{
B <- c(1, -1, 1.5, 1.5, rep(0, 6), 2, 0, 2, 0)
group <- c(rep('M1', 10), rep('M2', 10))
regressors <- matrix(rnorm(6*20), ncol = 6)
X  <- model.matrix(~group + group:regressors - 1)
y <- X\%*\%B + rnorm(20)
y <- scale(y)
mod <- fl2(y, regressors, group)
colors <- c(rep("grey",2), rep('green',2),rep('black', 6), rep(c("orange","blue"), 2), 'darkgreen', rep('yellow',3), rep('purple',2))
matplot(mod$lambda ,t(mod$beta),type='l',col=colors)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VariSel$new()}}
\item \href{#method-estime}{\code{VariSel$estime()}}
\item \href{#method-predict}{\code{VariSel$predict()}}
\item \href{#method-get_coef}{\code{VariSel$get_coef()}}
\item \href{#method-plot_path}{\code{VariSel$plot_path()}}
\item \href{#method-plot_error}{\code{VariSel$plot_error()}}
\item \href{#method-plot_BIC}{\code{VariSel$plot_BIC()}}
\item \href{#method-plot_coef}{\code{VariSel$plot_coef()}}
\item \href{#method-ROC}{\code{VariSel$ROC()}}
\item \href{#method-plot_anime}{\code{VariSel$plot_anime()}}
\item \href{#method-plot_cv}{\code{VariSel$plot_cv()}}
\item \href{#method-clone}{\code{VariSel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariSel$new(
  x,
  y,
  univ = TRUE,
  Sigma_12inv = diag(1, ncol(as.data.frame(y))),
  sepy = "__",
  sepx = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estime"></a>}}
\subsection{Method \code{estime()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariSel$estime()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariSel$predict(new_x, lambda = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lambda}}{if the user wants to use it owns values of lambdas}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_coef"></a>}}
\subsection{Method \code{get_coef()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariSel$get_coef()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_path"></a>}}
\subsection{Method \code{plot_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariSel$plot_path(type = "first", nb = 6)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_error"></a>}}
\subsection{Method \code{plot_error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariSel$plot_error(print = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_BIC"></a>}}
\subsection{Method \code{plot_BIC()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariSel$plot_BIC(print = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_coef"></a>}}
\subsection{Method \code{plot_coef()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariSel$plot_coef(tresh, sel = colnames(private$y))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ROC"></a>}}
\subsection{Method \code{ROC()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariSel$ROC(b)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_anime"></a>}}
\subsection{Method \code{plot_anime()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariSel$plot_anime(name_pos = NULL, num_max = 30)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_cv"></a>}}
\subsection{Method \code{plot_cv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariSel$plot_cv(sel = colnames(private$y), s = "lambda.min")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariSel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
