% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapeau.R
\name{compar_type}
\alias{compar_type}
\title{Title}
\usage{
compar_type(X, Y, types, Sigma_12inv = diag(1, ncol(as.data.frame(Y))),
  group = NULL, a = 1, times = 10, sep = "\\\\.")
}
\arguments{
\item{times}{}
}
\description{
Title
}
