% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fus2.R
\name{mod_fus2_univ}
\alias{mod_fus2_univ}
\title{Title}
\value{

}
\description{
Title

Title
}
\section{Super classes}{
\code{\link[VariSel:VariSel]{VariSel::VariSel}} -> \code{\link[VariSel:mod_lasso]{VariSel::mod_lasso}} -> \code{mod_fus2_univ}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{mod_fus2_univ$new()}}
\item \href{#method-predict}{\code{mod_fus2_univ$predict()}}
\item \href{#method-clone}{\code{mod_fus2_univ$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="get_coef">}\href{../../VariSel/html/VariSel.html#method-get_coef}{\code{VariSel::VariSel$get_coef()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="plot_anime">}\href{../../VariSel/html/VariSel.html#method-plot_anime}{\code{VariSel::VariSel$plot_anime()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="plot_BIC">}\href{../../VariSel/html/VariSel.html#method-plot_BIC}{\code{VariSel::VariSel$plot_BIC()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="plot_coef">}\href{../../VariSel/html/VariSel.html#method-plot_coef}{\code{VariSel::VariSel$plot_coef()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="plot_cv">}\href{../../VariSel/html/VariSel.html#method-plot_cv}{\code{VariSel::VariSel$plot_cv()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="plot_error">}\href{../../VariSel/html/VariSel.html#method-plot_error}{\code{VariSel::VariSel$plot_error()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="ROC">}\href{../../VariSel/html/VariSel.html#method-ROC}{\code{VariSel::VariSel$ROC()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="mod_lasso" data-id="estime">}\href{../../VariSel/html/mod_lasso.html#method-estime}{\code{VariSel::mod_lasso$estime()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="mod_lasso" data-id="plot_path">}\href{../../VariSel/html/mod_lasso.html#method-plot_path}{\code{VariSel::mod_lasso$plot_path()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="mod_lasso" data-id="sel_cv">}\href{../../VariSel/html/mod_lasso.html#method-sel_cv}{\code{VariSel::mod_lasso$sel_cv()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="mod_lasso" data-id="sel_stabs">}\href{../../VariSel/html/mod_lasso.html#method-sel_stabs}{\code{VariSel::mod_lasso$sel_stabs()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_fus2_univ$new(X, Y, group, a = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_fus2_univ$predict(new_x, new_group, lambda = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_fus2_univ$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
