% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_fused.R
\name{mod_fused_multi}
\alias{mod_fused_multi}
\title{Description of the function}
\description{
Description of the function

Description of the function
}
\section{Super classes}{
\code{\link[VariSel:VariSel]{VariSel::VariSel}} -> \code{\link[VariSel:mod_fused]{VariSel::mod_fused}} -> \code{mod_fused_multi}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{mod_fused_multi$new()}}
\item \href{#method-clone}{\code{mod_fused_multi$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="plot_anime">}\href{../../VariSel/html/VariSel.html#method-plot_anime}{\code{VariSel::VariSel$plot_anime()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="plot_BIC">}\href{../../VariSel/html/VariSel.html#method-plot_BIC}{\code{VariSel::VariSel$plot_BIC()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="plot_coef">}\href{../../VariSel/html/VariSel.html#method-plot_coef}{\code{VariSel::VariSel$plot_coef()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="plot_cv">}\href{../../VariSel/html/VariSel.html#method-plot_cv}{\code{VariSel::VariSel$plot_cv()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="plot_error">}\href{../../VariSel/html/VariSel.html#method-plot_error}{\code{VariSel::VariSel$plot_error()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="plot_path">}\href{../../VariSel/html/VariSel.html#method-plot_path}{\code{VariSel::VariSel$plot_path()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="predict">}\href{../../VariSel/html/VariSel.html#method-predict}{\code{VariSel::VariSel$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="VariSel" data-id="ROC">}\href{../../VariSel/html/VariSel.html#method-ROC}{\code{VariSel::VariSel$ROC()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="mod_fused" data-id="estime">}\href{../../VariSel/html/mod_fused.html#method-estime}{\code{VariSel::mod_fused$estime()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="VariSel" data-topic="mod_fused" data-id="get_coef">}\href{../../VariSel/html/mod_fused.html#method-get_coef}{\code{VariSel::mod_fused$get_coef()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_fused_multi$new(
  x,
  y,
  Sigma_12inv = diag(1, ncol(as.data.frame(y))),
  univ = FALSE,
  sepx = "\\\\."
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_fused_multi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
