% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chap_varisel.R
\name{train_VariSel}
\alias{train_VariSel}
\title{Title}
\usage{
train_VariSel(X, Y, type, sepx, reg, group, lambda = NULL,
  lambda1 = NULL, lambda2 = NULL)
}
\arguments{
\item{lambda2}{}
}
\description{
Title
}
